flatten filter config to the minimum needed
