+++ docs/reference/filters.md
@@ -817,13 +817,10 @@ skipper -enable-apimonitoring -enable-prometheus-metrics
 ```
 
 Endpoints can be monitored using the `apimonitoring` function in the route. It accepts a JSON object.
-* `apis`: A list of API being monitored by the filter.
-    * `application_id`: An application could be offering more than one API. Specify the application's ID here.
-    * `id`: The identifier of the API. For the moment, free text (is not yet used to fetch the API specification
-      in the API Repository). If not provided, the host name will be used to identify the API.
-    * `path_templates`: An endpoint path _template_, given in the OpenAPI format. Serves for grouping parametrized paths
-      together. Example: `/foo/1` and `/foo/2` should be monitored as the same endpoint, then provide: `PathPat: /foo/{foo-id}`.
-      It accepts both `{foo-id}` and `:foo-id` formats for the variable parts, but are all normalized to `:foo-id`.
+* `application_id`: An application could be offering more than one API. Specify the application's ID here.
+* `path_templates`: An endpoint path _template_, given in the OpenAPI format. Serves for grouping parametrized paths
+  together. Example: `/foo/1` and `/foo/2` should be monitored as the same endpoint, then provide: `PathPat: /foo/{foo-id}`.
+  It accepts both `{foo-id}` and `:foo-id` formats for the variable parts, but are all normalized to `:foo-id`.
 * `verbose` (default: `false`): An optional parameter making the filter log more detail about its operation.
   It is bypassed by the `--apimonitoring-verbose` switch when specified.
 
@@ -831,31 +828,20 @@ Example:
 
 ```
 apimonitoring(`{
-  "apis": [
-    {
-      "application_id": "my_app",
-      "id": "orders_api",
-      "path_templates": [
-        "foo/orders",
-        "foo/orders/:order-id",
-        "foo/orders/:order-id/order-items/{order-item-id}"
-      ]
-    },
-    {
-      "id": "customers_api",
-      "application_id": "my_app",
-      "path_templates": [
-        "/foo/customers/",
-        "/foo/customers/{customer-id}/"
-      ]
-    }
+  "application_id": "my_app",
+  "path_templates": [
+    "foo/orders",
+    "foo/orders/:order-id",
+    "foo/orders/:order-id/order-items/{order-item-id}"
+    "/foo/customers/",
+    "/foo/customers/{customer-id}/"
   ]
 }`)
 ```
 
 That would monitor metrics like:
-* `my_app.orders_api.GET.foo/orders/:order-id`
-* `my_app.orders_api.POST.foo/orders`
-* `my_app.orders_api.GET.foo/orders/:order-id/order-items/:order-item-id`
-* `my_app.customers_api.POST.foo/customers`
-* `my_app.customers_api.DELETE.foo/customers/:customer-id`
+* `api-mon.my_app.GET.foo/orders/:order-id`
+* `api-mon.my_app.POST.foo/orders`
+* `api-mon.my_app.GET.foo/orders/:order-id/order-items/:order-item-id`
+* `api-mon.my_app.POST.foo/customers`
+* `api-mon.my_app.DELETE.foo/customers/:customer-id`
