+++ filters/apimonitoring/config.go
@@ -1,12 +1,7 @@
 package apimonitoring
 
 type filterConfig struct {
-	Apis    []*filterConfigApi `json:"apis"`
-	Verbose bool               `json:"verbose"`
-}
-
-type filterConfigApi struct {
-	Id            string   `json:"id"`
+	Verbose       bool     `json:"verbose"`
 	ApplicationId string   `json:"application_id"`
 	PathTemplates []string `json:"path_templates"`
 }
