+++ filters/apimonitoring/filter_test.go
@@ -1,0 +1,173 @@
+package apimonitoring
+
+import (
+	"bytes"
+	"github.com/stretchr/testify/assert"
+	"github.com/zalando/skipper/filters"
+	"github.com/zalando/skipper/filters/filtertest"
+	"github.com/zalando/skipper/metrics/metricstest"
+	"io/ioutil"
+	"net/http"
+	"testing"
+)
+
+func createFilterForTest() (filters.Filter, error) {
+	spec := apiMonitoringFilterSpec{}
+	args := []interface{}{`{
+		"application_id": "my_app",
+		"path_templates": [
+			"foo/orders",
+			"foo/orders/:order-id",
+			"foo/orders/:order-id/order-items/{order-item-id}",
+			"/foo/customers/",
+			"/foo/customers/{customer-id}/"
+		]
+	}`}
+	return spec.CreateFilter(args)
+}
+
+func testWithFilter(
+	t *testing.T,
+	filterCreate func() (filters.Filter, error),
+	method string,
+	url string,
+	reqBody string,
+	resStatus int,
+	resBody string,
+	expect func(m *metricstest.MockMetrics, reqBodyLen int64, resBodyLen int64),
+) {
+	filter, err := filterCreate()
+	assert.NoError(t, err)
+
+	metricsMock := new(metricstest.MockMetrics)
+
+	req, err := http.NewRequest(method, url, bytes.NewBufferString(reqBody))
+	if err != nil {
+		t.Error(err)
+	}
+	ctx := &filtertest.Context{
+		FRequest: req,
+		FResponse: &http.Response{
+			StatusCode: resStatus,
+			Body:       ioutil.NopCloser(bytes.NewBufferString(resBody)),
+		},
+		FStateBag: make(map[string]interface{}),
+		FMetrics:  metricsMock,
+	}
+	filter.Request(ctx)
+	filter.Response(ctx)
+
+	expect(
+		metricsMock,
+		int64(len(reqBody)),
+		0, // int64(len(resBody)), // todo Restore after understanding why `response.ContentLength` returns always 0
+	)
+}
+
+func Test_Filter_NoPathPattern(t *testing.T) {
+	testWithFilter(
+		t,
+		createFilterForTest,
+		"GET",
+		"https://www.example.org/a/b/c",
+		"",
+		200,
+		"",
+		func(m *metricstest.MockMetrics, reqBodyLen int64, resBodyLen int64) {
+			// no path matching: no tracking
+			assert.Empty(t, m.Counters)
+			assert.Empty(t, m.Measures)
+		})
+}
+
+func Test_Filter_PathPatternNoVariablePart(t *testing.T) {
+	testWithFilter(
+		t,
+		createFilterForTest,
+		"POST",
+		"https://www.example.org/foo/orders",
+		"asd",
+		400,
+		"qwerty",
+		func(m *metricstest.MockMetrics, reqBodyLen int64, resBodyLen int64) {
+			assert.Equal(t,
+				map[string]int64{
+					"api-mon.my_app.POST.foo/orders.http_count":    1,
+					"api-mon.my_app.POST.foo/orders.http400_count": 1,
+					"api-mon.my_app.POST.foo/orders.req_size_sum":  reqBodyLen,
+					"api-mon.my_app.POST.foo/orders.resp_size_sum": resBodyLen,
+				},
+				m.Counters,
+			)
+			assert.Contains(t, m.Measures, "api-mon.my_app.POST.foo/orders.latency")
+		})
+}
+
+func Test_Filter_PathPatternWithVariablePart(t *testing.T) {
+	testWithFilter(
+		t,
+		createFilterForTest,
+		"POST",
+		"https://www.example.org/foo/orders/1234",
+		"asd",
+		400,
+		"qwerty",
+		func(m *metricstest.MockMetrics, reqBodyLen int64, resBodyLen int64) {
+			assert.Equal(t,
+				map[string]int64{
+					"api-mon.my_app.POST.foo/orders/:order-id.http_count":    1,
+					"api-mon.my_app.POST.foo/orders/:order-id.http400_count": 1,
+					"api-mon.my_app.POST.foo/orders/:order-id.req_size_sum":  reqBodyLen,
+					"api-mon.my_app.POST.foo/orders/:order-id.resp_size_sum": resBodyLen,
+				},
+				m.Counters,
+			)
+			assert.Contains(t, m.Measures, "api-mon.my_app.POST.foo/orders/:order-id.latency")
+		})
+}
+
+func Test_Filter_PathPatternWithMultipleVariablePart(t *testing.T) {
+	testWithFilter(
+		t,
+		createFilterForTest,
+		"POST",
+		"https://www.example.org/foo/orders/1234/order-items/123",
+		"asd",
+		400,
+		"qwerty",
+		func(m *metricstest.MockMetrics, reqBodyLen int64, resBodyLen int64) {
+			assert.Equal(t,
+				map[string]int64{
+					"api-mon.my_app.POST.foo/orders/:order-id/order-items/:order-item-id.http_count":    1,
+					"api-mon.my_app.POST.foo/orders/:order-id/order-items/:order-item-id.http400_count": 1,
+					"api-mon.my_app.POST.foo/orders/:order-id/order-items/:order-item-id.req_size_sum":  reqBodyLen,
+					"api-mon.my_app.POST.foo/orders/:order-id/order-items/:order-item-id.resp_size_sum": resBodyLen,
+				},
+				m.Counters,
+			)
+			assert.Contains(t, m.Measures, "api-mon.my_app.POST.foo/orders/:order-id/order-items/:order-item-id.latency")
+		})
+}
+
+func Test_Filter_PathPatternFromSecondConfiguredApi(t *testing.T) {
+	testWithFilter(
+		t,
+		createFilterForTest,
+		"POST",
+		"https://www.example.org/foo/customers/loremipsum",
+		"asd",
+		400,
+		"qwerty",
+		func(m *metricstest.MockMetrics, reqBodyLen int64, resBodyLen int64) {
+			assert.Equal(t,
+				map[string]int64{
+					"api-mon.my_app.POST.foo/customers/:customer-id.http_count":    1,
+					"api-mon.my_app.POST.foo/customers/:customer-id.http400_count": 1,
+					"api-mon.my_app.POST.foo/customers/:customer-id.req_size_sum":  reqBodyLen,
+					"api-mon.my_app.POST.foo/customers/:customer-id.resp_size_sum": resBodyLen,
+				},
+				m.Counters,
+			)
+			assert.Contains(t, m.Measures, "api-mon.my_app.POST.foo/customers/:customer-id.latency")
+		})
+}
